+++ libraries/DAC8571/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2024-04-18
+- implement more powerDown modi
+- update examples
+- add percentage wrappers
+- update keywords.txt
+- update readme.md 
+- minor edits
+
 ## [0.1.0] - 2024-04-16
 - initial version